requirements.txt
fastapi~=0.115.6
uvicorn~=0.32.1
pandas~=2.2.3
python-dotenv~=1.0.1
google-generativeai
pydantic~=2.10.2
protobuf~=5.29.1
httpx~=0.28.0